Add files via upload
This is the main program poorly written in Python with Tkinter (essentially my first try at python)  This script is designed to be used with a Raspberry Pi and a relay assigned to GPIO pin 26 and ground. The script will create a config file, then use the config file contents to create a beacon that runs every x seconds by calling the gerryk  https://github.com/gerryk/cwbeacon  beacon.py script with the appropriate arguments.